% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.largelle.R
\name{plot.largeLLE}
\alias{plot.largeLLE}
\title{Plot coefficients from a "largeLLE" object}
\usage{
\method{plot}{largeLLE}(x, alpha = 1, log.l = TRUE, ...)
}
\arguments{
\item{x}{Fitted \code{"largeLLE"} model.}

\item{alpha}{Controls alpha-blending, helpful when the number of covariates
is large.  Default is alpha=1.}

\item{log.l}{Should horizontal axis be on the log scale?  Default is TRUE.}

\item{\dots}{Other graphical parameters to \code{plot}}
}
\description{
Produce a plot of the coefficient paths for a fitted \code{\link{largeLLE}}
object (this is functionally unchanged from the biglasso function).
}
\examples{

## See examples in "largeLLE"

}
\seealso{
\code{\link{largeLLE}}, \code{\link{cv.largelle}}
}
\author{
Jason Orender, Yaohui Zeng and Patrick Breheny

Maintainer: Jason Orender <joren001@odu.edu>
}
\keyword{models}
\keyword{regression}
