% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mlargelle.R
\name{plot.mlargelle}
\alias{plot.mlargelle}
\title{Plot coefficients from a "mlargelle" object}
\usage{
\method{plot}{mlargelle}(x, alpha = 1, log.l = TRUE, norm.beta = TRUE, ...)
}
\arguments{
\item{x}{Fitted \code{"mlargelle"} model.}

\item{alpha}{Controls alpha-blending, helpful when the number of covariates
is large.  Default is alpha=1.}

\item{log.l}{Should horizontal axis be on the log scale?  Default is TRUE.}

\item{norm.beta}{Should the vertical axis be the l2 norm of coefficients for each variable?
Default is TRUE. If False, the vertical axis is the coefficients.}

\item{\dots}{Other graphical parameters to \code{plot}}
}
\description{
Produce a plot of the coefficient paths for a fitted multiple responses
\code{mlargelle} object.
}
\examples{

## See examples in "largeLLE"

}
\seealso{
\code{\link{largeLLE}},
}
\author{
Jason Orender, Chuyi Wang

Maintainer: Jason Orender <joren001@odu.edu>
}
\keyword{models}
\keyword{regression}
