% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.R
\name{predict.cv.largelle}
\alias{predict.cv.largelle}
\alias{coef.cv.largelle}
\title{Model predictions based on a fitted \code{\link{cv.largelle}} object}
\usage{
\method{predict}{cv.largelle}(
  object,
  X,
  row.idx = 1:nrow(X),
  type = c("link", "response", "class", "coefficients", "vars", "nvars"),
  lambda = object$lambda.min,
  which = object$min,
  ...
)

\method{coef}{cv.largelle}(object, lambda = object$lambda.min, which = object$min, ...)
}
\arguments{
\item{object}{A fitted \code{"cv.largelle"} model object.}

\item{X}{Matrix of values at which predictions are to be made. It must be a
\code{\link[bigmemory]{big.matrix}} object. Not used for
\code{type="coefficients"}.}

\item{row.idx}{Similar to that in \code{\link[largeLLE]{largeLLE}}, it's a
vector of the row indices of \code{X} that used for the prediction.
\code{1:nrow(X)} by default.}

\item{type}{Type of prediction: \code{"link"} returns the linear predictors;
\code{"response"} gives the fitted values; \code{"class"} returns the
binomial outcome with the highest probability; \code{"coefficients"} returns
the coefficients; \code{"vars"} returns a list containing the indices and
names of the nonzero variables at each value of \code{lambda};
\code{"nvars"} returns the number of nonzero coefficients at each value of
\code{lambda}.}

\item{lambda}{Values of the regularization parameter \code{lambda} at which
predictions are requested.  The default value is the one corresponding to
the minimum cross-validation error. Accepted values are also the strings
"lambda.min" (\code{lambda} of minimum cross-validation error) and
"lambda.1se" (Largest value of \code{lambda} for which the cross-validation
error was at most one standard error larger than the minimum.).}

\item{which}{Indices of the penalty parameter \code{lambda} at which
predictions are requested. The default value is the index of lambda
corresponding to lambda.min.  Note: this is overridden if \code{lambda} is
specified.}

\item{\dots}{Not used.}
}
\value{
The object returned depends on \code{type}.
}
\description{
Extract predictions from a fitted \code{\link{cv.largelle}} object.
}
\examples{
\dontrun{
## predict.cv.largelle
data(colon)
X <- colon$X
y <- colon$y
X.bm <- as.big.matrix(X, backingfile = "")
fit <- largeLLE(X.bm, y, penalty = 'lasso', family = "binomial")
cvfit <- cv.largelle(X.bm, y, penalty = 'lasso', family = "binomial", seed = 1234, ncores = 2)
coef <- coef(cvfit)
coef[which(coef != 0)]
predict(cvfit, X.bm, type = "response")
predict(cvfit, X.bm, type = "link")
predict(cvfit, X.bm, type = "class")
predict(cvfit, X.bm, lambda = "lambda.1se")
}
}
\seealso{
\code{\link{largeLLE}}, \code{\link{cv.largelle}}
}
\author{
Yaohui Zeng and Patrick Breheny

Maintainer: Yaohui Zeng <yaohui.zeng@gmail.com>
}
\keyword{models}
\keyword{regression}
