% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cv.largelle.R
\name{summary.cv.largelle}
\alias{summary.cv.largelle}
\alias{print.summary.cv.largelle}
\title{Summarizing inferences based on cross-validation}
\usage{
\method{summary}{cv.largelle}(object, ...)

\method{print}{summary.cv.largelle}(x, digits, ...)
}
\arguments{
\item{object}{A \code{cv.largelle} object.}

\item{\dots}{Further arguments passed to or from other methods.}

\item{x}{A \code{"summary.cv.largelle"} object.}

\item{digits}{Number of digits past the decimal point to print out.  Can be
a vector specifying different display digits for each of the five
non-integer printed values.}
}
\value{
\code{summary.cv.largelle} produces an object with S3 class
\code{"summary.cv.largelle"}.  The class has its own print method and contains
the following list elements: \item{penalty}{The penalty used by
\code{largeLLE}.} \item{model}{Either \code{"linear"} or \code{"logistic"},
depending on the \code{family} option in \code{largeLLE}.} \item{n}{Number
of observations} \item{p}{Number of regression coefficients (not including
the intercept).} \item{min}{The index of \code{lambda} with the smallest
cross-validation error.} \item{lambda}{The sequence of \code{lambda} values
used by \code{cv.largelle}.} \item{cve}{Cross-validation error (deviance).}
\item{r.squared}{Proportion of variance explained by the model, as estimated
by cross-validation.} \item{snr}{Signal to noise ratio, as estimated by
cross-validation.} \item{sigma}{For linear regression models, the scale
parameter estimate.} \item{pe}{For logistic regression models, the
prediction error (misclassification error).}
}
\description{
Summary method for \code{cv.largelle} objects.
}
\examples{
## See examples in "cv.largelle" and "largeLLE-package"
}
\seealso{
\code{\link{largeLLE}}, \code{\link{cv.largelle}},
\code{\link{plot.cv.largelle}}
}
\author{
Yaohui Zeng and Patrick Breheny

Maintainer: Jason Orender <joren001@odu.edu>
}
\keyword{models}
\keyword{regression}
